// src/app/models/player.ts
export interface Player {
  playerId: string;
  playerName: string;
  dob: string;
  mlbDebutDate: string;
  birthCity: string;
  birthStateProvince: string;
  birthCountry: string;
  heightInches: number;
  weight: number;
  primaryPositionCode: string;
  primaryPositionName: string;
  playerForTestSetAndFuturePreds: boolean;

}